requirements.txt
jupyter
pandas
scikit-learn
pillow
matplotlib
torchsummary
tqdm
opencv-python